[Modify] Layout adjustment due to folder grouping addition
