+++ src/app/layout.tsx
@@ -1,0 +1,27 @@
+import type { Metadata } from 'next';
+import { Inter } from 'next/font/google';
+import '@/style/globals.css';
+import { Providers } from '@/components/Providers';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+  title: 'Bottle Note',
+  description: '위스키 라이프를 기록하다, 보틀 노트.',
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={inter.className} suppressHydrationWarning>
+        <Providers>
+          <body className={inter.className}>{children}</body>
+        </Providers>
+      </body>
+    </html>
+  );
+}
